Code/requirements.txt
praw
python-dotenv
pandas
requests
tqdm
